[#5289] Fix error when a journal with a question is destroyed by clearing the notes

+++ lib/question_journal_hooks.rb
@@ -22,8 +22,10 @@ class QuestionJournalHooks < QuestionHooksBase
     journal = context[:journal]
     params = context[:params]
 
+    # Handle destroying journals through the 'edit' action (done by clearing notes)
+    return '' if journal.destroyed?
+    
     if params[:question] && params[:question][:assigned_to]
-
       if journal.question && params[:question][:assigned_to].blank?
         # Wants to remove the question
         journal.question.destroy
